Implemented new System Time framework
#145 Moved all GPS source to its library lib-gps
