+++ h3-firmware-template/Rules.mk
@@ -109,6 +109,8 @@ ifneq ($(CONSOLE),)
 	DEFINES+=-D$(CONSOLE)
 endif
 
+DEFINES+=-D_TIME_STAMP_YEAR_=$(shell date  +"%Y") -D_TIME_STAMP_MONTH_=$(shell date  +"%-m") -D_TIME_STAMP_DAY_=$(shell date  +"%-d")
+
 # The variable for the firmware include directories
 INCDIRS+=../include $(wildcard ./include) $(wildcard ./*/include)
 INCDIRS:=$(addprefix -I,$(INCDIRS))
