+++ lib-gps/src/gpsconst.cpp
