+++ lib-gps/src/gpsstatic.cpp
