+++ lib-network/include/ntpclient.h
@@ -23,6 +23,10 @@
  * THE SOFTWARE.
  */
 
+/*
+ * PoC Code -> Do not use in production
+ */
+
 #ifndef NTPCLIENT_H_
 #define NTPCLIENT_H_
 
@@ -32,9 +36,8 @@
 #include "ntp.h"
 
 enum class NtpClientStatus {
-	INIT,
 	IDLE,
-	STOPPED,
+	FAILED,
 	WAITING
 };
 
@@ -50,7 +53,7 @@ class NtpClient {
 public:
 	NtpClient(uint32_t nServerIp = 0);
 
-	void Init();
+	void Start();
 	void Run();
 
 	void Print();
@@ -69,17 +72,37 @@ public:
 
 private:
 	void SetUtcOffset(float fUtcOffset);
+	void GetTimeNtpFormat(uint32_t &nSeconds, uint32_t &nFraction);
+	void Send();
+	bool Receive();
+
+	struct TimeStamp {
+		uint32_t nSeconds;
+		uint32_t nFraction;
+	};
+
+	void Difference(const struct TimeStamp *Start, const struct TimeStamp *Stop, int32_t &nDiffSeconds, uint32_t &nDiffFraction);
+	int SetTimeOfDay();
+
+	void PrintNtpTime(const char *pText, const struct TimeStamp *pNtpTime);
 
 private:
 	uint32_t m_nServerIp;
 	int32_t m_nUtcOffset;
 	int32_t m_nHandle{-1};
-	NtpClientStatus m_tStatus{NtpClientStatus::STOPPED};
+	NtpClientStatus m_tStatus{NtpClientStatus::IDLE};
 	struct TNtpPacket m_Request;
 	struct TNtpPacket m_Reply;
-	time_t m_InitTime{0};
 	uint32_t m_MillisRequest{0};
 	uint32_t m_MillisLastPoll{0};
+
+	struct TimeStamp T1{0,0};	// time request sent by client
+	struct TimeStamp T2{0,0};	// time request received by server
+	struct TimeStamp T3{0,0};	// time reply sent by server
+	struct TimeStamp T4{0,0};	// time reply received by client
+
+	int32_t m_nOffsetSeconds{0};
+	uint32_t m_nOffsetMicros{0};
 
 	NtpClientDisplay *m_pNtpClientDisplay = nullptr;
