+++ lib-network/src/ntpclient.cpp
@@ -23,9 +23,14 @@
  * THE SOFTWARE.
  */
 
+/*
+ * PoC Code -> Do not use in production
+ */
+
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
+#include <sys/time.h>
 #include <cassert>
 
 #include "ntpclient.h"
@@ -44,20 +49,36 @@ extern "C" {
 
 #include "debug.h"
 
-#define RETRIES			3
-#define TIMEOUT_MILLIS	3000 	// 3 seconds
-#define POLL_SECONDS	1024	// 2ˆ10
+static constexpr auto RETRIES = 3;
+static constexpr auto TIMEOUT_MILLIS = 3000; 	// 3 seconds
+static constexpr auto POLL_SECONDS = 1024;		// 2ˆ10
+static constexpr auto JAN_1970 = 0x83aa7e80; 	// 2208988800 1970 - 1900 in seconds
+
+/* How to multiply by 4294.967296 quickly (and not quite exactly)
+ * without using floating point or greater than 32-bit integers.
+ * If you want to fix the last 12 microseconds of error, add in
+ * (2911*(x))>>28)
+ */
+#define NTPFRAC(x) ( 4294U*static_cast<uint32_t>(x) + ( (1981U*static_cast<uint32_t>(x))>>11 ) +  ((2911U*static_cast<uint32_t>(x))>>28) )
+
+/* The reverse of the above, needed if we want to set our microsecond
+ * clock (via clock_settime) based on the incoming time in NTP format.
+ * Basically exact.
+ */
+#define USEC(x) ( ( (x) >> 12 ) - 759 * ( ( ( (x) >> 10 ) + 32768 ) >> 16 ) )
+
+/*
+Timestamp Name        ID   When Generated
+----------------------------------------------------------------
+Originate Timestamp   T1   time request sent by client
+Receive Timestamp     T2   time request received by server
+Transmit Timestamp    T3   time reply sent by server
+Destination Timestamp T4   time reply received by client
+*/
 
 NtpClient *NtpClient::s_pThis = nullptr;
 
-NtpClient::NtpClient(uint32_t nServerIp):
-	m_nServerIp(nServerIp),
-	m_nHandle(-1),
-	m_tStatus(NtpClientStatus::STOPPED),
-	m_InitTime(0),
-	m_MillisRequest(0),
-	m_MillisLastPoll(0)
-{
+NtpClient::NtpClient(uint32_t nServerIp): m_nServerIp(nServerIp) {
 	DEBUG_ENTRY
 	assert(s_pThis == nullptr);
 	s_pThis = this;
@@ -69,11 +90,11 @@ NtpClient::NtpClient(uint32_t nServerIp):
 	SetUtcOffset(Network::Get()->GetNtpUtcOffset());
 
 	memset(&m_Request, 0, sizeof m_Request);
+	memset(&m_Reply, 0, sizeof m_Reply);
 
 	m_Request.LiVnMode = NTP_VERSION | NTP_MODE_CLIENT;
 	m_Request.Poll = 10; // Poll: 1024 seconds
-
-	memset(&m_Reply, 0, sizeof m_Reply);
+	m_Request.ReferenceID = ('A' << 0) | ('V' << 8) | ('S' << 16);
 
 	DEBUG_EXIT
 }
@@ -83,10 +104,113 @@ void NtpClient::SetUtcOffset(float fUtcOffset) {
 	m_nUtcOffset = Utc::Validate(fUtcOffset);
 }
 
-void NtpClient::Init() {
+/*
+ * Seconds and Fractions since 01.01.1900
+ */
+void NtpClient::GetTimeNtpFormat(uint32_t &nSeconds, uint32_t &nFraction) {
+	struct timeval now;
+	gettimeofday(&now, nullptr);
+	nSeconds = static_cast<uint32_t>(now.tv_sec - m_nUtcOffset) + JAN_1970;
+	nFraction = NTPFRAC(now.tv_usec);
+}
+
+void NtpClient::Send() {
+	GetTimeNtpFormat(T1.nSeconds, T1.nFraction);
+
+	m_Request.OriginTimestamp_s = __builtin_bswap32(T1.nSeconds);
+	m_Request.OriginTimestamp_f = __builtin_bswap32(T1.nFraction);
+
+	Network::Get()->SendTo(m_nHandle, &m_Request, sizeof m_Request, m_nServerIp, NTP_UDP_PORT);
+}
+
+bool NtpClient::Receive() {
+	uint32_t nFromIp;
+	uint16_t nFromPort;
+
+	const uint16_t nBytesReceived = Network::Get()->RecvFrom(m_nHandle, &m_Reply, sizeof m_Reply, &nFromIp, &nFromPort);
+
+	if (__builtin_expect((nBytesReceived != sizeof m_Reply), 1)) {
+		return false;
+	}
+
+	GetTimeNtpFormat(T4.nSeconds, T4.nFraction);
+
+	if (__builtin_expect((nFromIp != m_nServerIp), 0)) {
+		DEBUG_PUTS("nFromIp != m_nServerIp");
+		return false;
+	}
+
+	T2.nSeconds = __builtin_bswap32(m_Reply.ReceiveTimestamp_s);
+	T2.nFraction = __builtin_bswap32(m_Reply.ReceiveTimestamp_f);
+
+	T3.nSeconds = __builtin_bswap32(m_Reply.TransmitTimestamp_s);
+	T3.nFraction = __builtin_bswap32(m_Reply.TransmitTimestamp_f);
+
+	PrintNtpTime("Originate", &T1);
+	PrintNtpTime("Receive", &T2);
+	PrintNtpTime("Transmit", &T3);
+	PrintNtpTime("Destination", &T4);
+
+	return true;
+}
+
+void NtpClient::Difference(const struct TimeStamp *Start, const struct TimeStamp *Stop, int32_t &nDiffSeconds, uint32_t &nDiffMicros) {
+	nDiffSeconds = static_cast<int32_t>(Stop->nSeconds - Start->nSeconds);
+	uint32_t nDiffFraction;
+
+	if (Stop->nFraction >= Start->nFraction) {
+		nDiffFraction = Stop->nFraction - Start->nFraction;
+	} else {
+		nDiffFraction = Start->nFraction - Stop->nFraction;
+		nDiffFraction = ~nDiffFraction;
+		nDiffSeconds -= 1;
+	}
+
+	nDiffMicros = USEC(nDiffFraction);
+
+	DEBUG_PRINTF("Seconds  %u - %u = %d", Stop->nSeconds, Start->nSeconds, nDiffSeconds);
+	DEBUG_PRINTF("Micros   %u - %u = %u", USEC(Stop->nFraction), USEC(Start->nFraction), nDiffMicros);
+}
+
+int NtpClient::SetTimeOfDay() {
+	int32_t nDiffSeconds1, nDiffSeconds2 ;
+	uint32_t nDiffFraction1, nDiffFraction2;
+
+	Difference(&T1, &T2, nDiffSeconds1, nDiffFraction1);
+	Difference(&T4, &T3, nDiffSeconds2, nDiffFraction2);
+
+	m_nOffsetSeconds = nDiffSeconds1 + nDiffSeconds2;
+	m_nOffsetMicros = nDiffFraction1 + nDiffFraction2;
+
+	if (m_nOffsetMicros >= 1E6) {
+		m_nOffsetMicros -= 1E6	;
+		m_nOffsetSeconds += 1;
+	}
+
+	m_nOffsetSeconds /= 2;
+	m_nOffsetMicros /= 2;
+
+	struct timeval tv;
+
+	tv.tv_sec =	static_cast<time_t>(T4.nSeconds - JAN_1970)  + m_nOffsetSeconds + m_nUtcOffset;
+	tv.tv_usec = static_cast<suseconds_t>(static_cast<int32_t>(USEC(T4.nFraction)) + static_cast<int32_t>(m_nOffsetMicros));
+
+	if (tv.tv_usec >= 1E6) {
+		tv.tv_sec++;
+		tv.tv_usec = 1E6 - tv.tv_usec;
+	}
+
+	DEBUG_PRINTF("(%u, %u) %s ", tv.tv_sec, tv.tv_usec , tv.tv_usec  >= 1E6 ? "!" : "");
+	DEBUG_PRINTF("%d %u",m_nOffsetSeconds, m_nOffsetMicros);
+
+	return settimeofday(&tv, nullptr);
+}
+
+void NtpClient::Start() {
 	DEBUG_ENTRY
 
 	if (m_nServerIp == 0) {
+		m_tStatus = NtpClientStatus::FAILED;
 		DEBUG_EXIT
 		return;
 	}
@@ -95,20 +219,16 @@ void NtpClient::Init() {
 	assert(m_nHandle != -1);
 
 	if (m_pNtpClientDisplay != nullptr) {
-		m_pNtpClientDisplay->ShowNtpClientStatus(NtpClientStatus::INIT);
+		m_pNtpClientDisplay->ShowNtpClientStatus(NtpClientStatus::IDLE);
 	}
 
 	const uint32_t nNow = Hardware::Get()->Millis();
 	uint32_t nRetries;
-	uint16_t nBytesReceived;
 
 	for (nRetries = 0; nRetries < RETRIES; nRetries++) {
-		Network::Get()->SendTo(m_nHandle, &m_Request, sizeof m_Request, m_nServerIp, NTP_UDP_PORT);
+		Send();
 
-		uint32_t nFromIp;
-		uint16_t nFromPort;
-
-		while ((nBytesReceived = Network::Get()->RecvFrom(m_nHandle, &m_Reply, sizeof m_Reply, &nFromIp, &nFromPort)) == 0) {
+		while (!Receive()) {
 #if defined (H3)
 			net_handle();
 #endif
@@ -117,51 +237,38 @@ void NtpClient::Init() {
 			}
 		}
 
-		if (nBytesReceived == sizeof m_Reply) {
-			if (__builtin_expect((nFromIp != m_nServerIp), 0)) {
-				DEBUG_PUTS("nFromIp != m_nServerIp");
-				continue;
-			}
-
-			debug_dump(&m_Reply, sizeof m_Reply);
-
-			if ((m_Reply.LiVnMode & NTP_MODE_SERVER) == NTP_MODE_SERVER) {
-				m_InitTime = ((__builtin_bswap32(m_Reply.ReceiveTimestamp_s) - NTP_TIMESTAMP_DELTA + m_nUtcOffset));
-				DEBUG_PRINTF("m_InitTime=%u", static_cast<unsigned>(m_InitTime));
-
-				struct tm *pLocalTime = localtime(&m_InitTime);
-
-				DEBUG_PRINTF("%.4d/%.2d/%.2d %.2d:%.2d:%.2d", pLocalTime->tm_year, pLocalTime->tm_mon, pLocalTime->tm_mday, pLocalTime->tm_hour, pLocalTime->tm_min, pLocalTime->tm_sec);
-
-				if (Hardware::Get()->SetTime(pLocalTime)) {
-					m_MillisLastPoll = Hardware::Get()->Millis();
-					m_tStatus = NtpClientStatus::IDLE;
-				}
+		if ((m_Reply.LiVnMode & NTP_MODE_SERVER) == NTP_MODE_SERVER) {
+			if (SetTimeOfDay() == 0) {
+				m_tStatus = NtpClientStatus::IDLE;
 			} else {
-				DEBUG_PUTS("!>> Invalid reply <<!");
+				// Error
 			}
-			break;
+		} else {
+			DEBUG_PUTS("!>> Invalid reply <<!");
 		}
+
+		break;
 	}
 
-	if (m_tStatus == NtpClientStatus::STOPPED) {
+	m_MillisLastPoll = Hardware::Get()->Millis();
+
+	if (nRetries == RETRIES) {
+		m_tStatus = NtpClientStatus::FAILED;
+
 		if (m_pNtpClientDisplay != nullptr) {
-			m_pNtpClientDisplay->ShowNtpClientStatus(NtpClientStatus::STOPPED);
+			m_pNtpClientDisplay->ShowNtpClientStatus(NtpClientStatus::FAILED);
 		}
 	}
 
-	DEBUG_PRINTF("nBytesReceived=%d, nRetries=%d, m_tStatus=%d", nBytesReceived, nRetries, static_cast<int>(m_tStatus));
+
+	DEBUG_PRINTF("nRetries=%d, m_tStatus=%d", nRetries, static_cast<int>(m_tStatus));
 	DEBUG_EXIT
 }
 
 void NtpClient::Run() {
-	if (m_tStatus == NtpClientStatus::STOPPED) {
-		return;
-	}
-
-	if (m_tStatus == NtpClientStatus::IDLE) {
+	if ((m_tStatus == NtpClientStatus::IDLE) || (m_tStatus == NtpClientStatus::FAILED)) {
 		if (__builtin_expect(((Hardware::Get()->Millis() - m_MillisLastPoll) > (1000 * POLL_SECONDS)), 0)) {
-			Network::Get()->SendTo(m_nHandle, &m_Request, sizeof m_Request, m_nServerIp, NTP_UDP_PORT);
+			Send();
 			m_MillisRequest = Hardware::Get()->Millis();
 			m_tStatus = NtpClientStatus::WAITING;
 			DEBUG_PUTS("NtpClientStatus::WAITING");
@@ -171,39 +278,31 @@ void NtpClient::Run() {
 	}
 
 	if (m_tStatus == NtpClientStatus::WAITING) {
-		uint32_t nFromIp;
-		uint16_t nFromPort;
-
-		if ((Network::Get()->RecvFrom(m_nHandle, &m_Reply, sizeof m_Reply, &nFromIp, &nFromPort)) != sizeof m_Reply) {
+		if (!Receive()) {
 			if (__builtin_expect(((Hardware::Get()->Millis() - m_MillisRequest) > TIMEOUT_MILLIS), 0)) {
-				m_tStatus = NtpClientStatus::STOPPED;
-				if (m_tStatus == NtpClientStatus::STOPPED) {
-					if (m_pNtpClientDisplay != nullptr) {
-						m_pNtpClientDisplay->ShowNtpClientStatus(NtpClientStatus::STOPPED);
-					}
+				m_tStatus = NtpClientStatus::FAILED;
+
+				if (m_pNtpClientDisplay != nullptr) {
+					m_pNtpClientDisplay->ShowNtpClientStatus(NtpClientStatus::FAILED);
 				}
-				DEBUG_PUTS("NtpClientStatus::STOPPED");
+				DEBUG_PUTS("NtpClientStatus::FAILED");
 			}
+
 			return;
 		}
-
-		if (__builtin_expect((nFromIp != m_nServerIp), 0)) {
-			DEBUG_PUTS("nFromIp != m_nServerIp");
-			return;
-		}
-
-		debug_dump(&m_Reply, sizeof m_Reply);
 
 		if (__builtin_expect(((m_Reply.LiVnMode & NTP_MODE_SERVER) == NTP_MODE_SERVER), 1)) {
-			const time_t nTime = ((__builtin_bswap32(m_Reply.ReceiveTimestamp_s) - NTP_TIMESTAMP_DELTA + m_nUtcOffset));
-			Hardware::Get()->SetSysTime(nTime);
 			m_MillisLastPoll = Hardware::Get()->Millis();
 
+			if (SetTimeOfDay() == 0) {
 #ifndef NDEBUG
-			DEBUG_PRINTF("nTime=%u", static_cast<unsigned>(nTime));
-			struct tm *pLocalTime = localtime(&nTime);
-			DEBUG_PRINTF("%.4d/%.2d/%.2d %.2d:%.2d:%.2d", pLocalTime->tm_year, pLocalTime->tm_mon, pLocalTime->tm_mday, pLocalTime->tm_hour, pLocalTime->tm_min, pLocalTime->tm_sec);
+				const time_t nTime = time(nullptr);
+				const struct tm *pLocalTime = localtime(&nTime);
+				DEBUG_PRINTF("%.4d/%.2d/%.2d %.2d:%.2d:%.2d", pLocalTime->tm_year, pLocalTime->tm_mon, pLocalTime->tm_mday, pLocalTime->tm_hour, pLocalTime->tm_min, pLocalTime->tm_sec);
 #endif
+			} else {
+				// Error
+			}
 		} else {
 			DEBUG_PUTS("!>> Invalid reply <<!");
 		}
@@ -213,15 +312,26 @@ void NtpClient::Run() {
 	}
 }
 
+void NtpClient::PrintNtpTime(__attribute__((unused)) const char *pText, __attribute__((unused)) const struct TimeStamp *pNtpTime) {
+#ifndef NDEBUG
+	const time_t nSeconds = static_cast<time_t>(pNtpTime->nSeconds - JAN_1970);
+	const struct tm *pTm = localtime(&nSeconds);
+	printf("%s %02d:%02d:%02d.%06d %04d [%u]\n", pText, pTm->tm_hour, pTm->tm_min,  pTm->tm_sec, USEC(pNtpTime->nFraction), pTm->tm_year + 1900, pNtpTime->nSeconds);
+#endif
+}
+
 void NtpClient::Print() {
 	printf("NTP v%d Client\n", NTP_VERSION >> 3);
 	if (m_nServerIp == 0) {
 		printf(" Not enabled\n");
 		return;
 	}
-	printf(" Server : " IPSTR "\n", IP2STR(m_nServerIp));
-	printf(" Port : %d\n", NTP_UDP_PORT);
-	printf(" Status : %d%c\n", static_cast<int>(m_tStatus), m_tStatus == NtpClientStatus::STOPPED ? '!' : ' ');
-	printf(" Time : %s", asctime(localtime(&m_InitTime)));
+	printf(" Server : " IPSTR ":%d\n", IP2STR(m_nServerIp), NTP_UDP_PORT);
+	printf(" Status : %d\n", static_cast<int>(m_tStatus));
 	printf(" UTC offset : %d (seconds)\n", m_nUtcOffset);
+	// Debug ONLY
+	PrintNtpTime("Originate", &T1);
+	PrintNtpTime("Receive", &T2);
+	PrintNtpTime("Transmit", &T3);
+	PrintNtpTime("Destination", &T4);
 }
