--- lib-tlc59711/.tracecompass/.project
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<projectDescription>
-	<name>.tracecompass-lib-tlc59711</name>
-	<comment></comment>
-	<projects>
-	</projects>
-	<buildSpec>
-	</buildSpec>
-	<natures>
-		<nature>org.eclipse.linuxtools.tmf.project.nature</nature>
-	</natures>
-</projectDescription>
