+++ opi_emac_artnet_monitor/firmware/main.cpp
@@ -102,8 +102,13 @@ void notmain(void) {
 
 	NtpClient ntpClient;
 	ntpClient.SetNtpClientDisplay(&displayUdfHandler);
-	ntpClient.Init();
+	ntpClient.Start();
 	ntpClient.Print();
+
+	if (ntpClient.GetStatus() != NtpClientStatus::FAILED) {
+		printf("Set RTC from System Clock\n");
+		HwClock::Get()->SysToHc();
+	}
 
 	display.TextStatus(ArtNetMsgConst::PARAMS, Display7SegmentMessage::INFO_NODE_PARMAMS, CONSOLE_YELLOW);
