+++ opi_emac_ltc_smpte/Makefile.H3
@@ -1,7 +1,7 @@
 #
-DEFINES = LTC_READER RDMNET_LLRP_ONLY ENABLE_SHELL NDEBUG
+DEFINES = LTC_READER RDMNET_LLRP_ONLY NDEBUG
 #
-LIBS = ltc tcnet artnet artnethandlers midi input osc ws28xxdisplay ws28xx rdmnet rdm rdmsensor rdmsubdevice
+LIBS = ltc tcnet artnet artnethandlers midi input osc ws28xxdisplay ws28xx rdmnet rdm rdmsensor rdmsubdevice gps
 #
 SRCDIR = firmware lib
